Refactor(Product and ProductImage): - Refactor ProductService interface to use ProductResponse and ProductImageResponse DTOs; update methods for better error handling and image management
- Enhance ProductImagesDialog component to support image deletion and improved image handling

- Modify Featured and OnSale components to handle different response structures for product lists

- Refactor SearchBar, ProductsPage, ProductsByCategoryPage, ProductsByBrandPage, and ProductPage components to accommodate new response formats

+++ pcshopbackend/pcshop/src/main/java/com/project/pcshop/responses/ProductResponse.java
@@ -24,7 +24,7 @@ public class ProductResponse {
 
     public static ProductResponse fromProduct(Product product) {
         return ProductResponse.builder()
-                .id(product.getId().longValue())
+                .id(product.getId())
                 .name(product.getName())
                 .price(product.getPrice())
                 .discount(product.getDiscount())
